+++ package.json
@@ -30,6 +30,7 @@
     "cjs": true
   },
   "dependencies": {
+    "@aquarius/regex": "^1.0.0",
     "@google-cloud/firestore": "^2.6.0",
     "@sentry/node": "5.9.0",
     "alphavantage": "^2.0.0",
