+++ packages/regex/.babelrc
@@ -1,0 +1,13 @@
+{
+  "presets": [
+    [
+      "env",
+      {
+        "targets": {
+          "node": "12.0.0"
+        }
+      }
+    ]
+  ],
+  "plugins": ["add-module-exports"]
+}
