+++ packages/regex/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@aquarius/regex",
+  "version": "1.0.0",
+  "description": "Regex matchers for Discord",
+  "author": "Ian Mitchell <ian.mitchel1@live.com>",
+  "homepage": "https://github.com/ianmitchell/aquarius",
+  "repository": "ianmitchell/aquarius",
+  "license": "MIT",
+  "scripts": {
+    "build": "babel src --out-dir lib --copy-files"
+  },
+  "keywords": [
+    "discord",
+    "discord.js",
+    "bot",
+    "regex",
+    "node"
+  ],
+  "main": "regex.js"
+}
