+++ packages/regex/src/regex.js
